Added configurable permaprop limit
Added a config item that allows the server admin to set a permaprop limit.

When loading permaprops, the script will load the file sequentially until it hits the limit and then stop.

As the script currently saves permaprops at server shutdown/map change, the permaprops that were not loaded are lost permanently.

1) Leave it as is. Permaprops above the count limit will disappear forever on restart.

2) Save permaprops to file immediately after the user runs rpa_togglesaved, instead of at server shutdown. Don't know the performance impact of this on live server TBH, could be none, could be a tiny lag spike.

3) On server shutdown, if the permaprop setup is different, save a new file with a timestamp as a way of not overwriting the old permaprop file. When loading permaprops next time, select the filename with the latest timestamp (with the lower limit), with the option for whoever administrates the server to manually restore the old backup.