Added dynamic permaprop limit
Permaprops are no longer "chopped off" during loading if the limit has been exceeded. Instead, they are all loaded in but it is impossible to add new permaprops, only remove.

rpa_togglesaved now tracks the number of permaprops as it adds and removes.

Fixed a typo in sh_items.lua

It's not ideal from a user perspective yet
If I use my rpa_togglesaved bind I don't want to accidentally de-permaprop a prop, possibly several in a row if I'm spamming it, and not be able to re-add it until props have been removed back down to the limit.